Resolved #382 - Adapted the demoApp to the new FlexibleItemDecoration and fixed some bugs around.

+++ flexible-adapter-app/src/main/java/eu/davidea/utils/Utils.java
@@ -54,7 +54,7 @@ public final class Utils {
 		return context.getResources().getDisplayMetrics();
 	}
 
-	public static float dpToPx(Context context, float dp) {
+	public static int dpToPx(Context context, float dp) {
 		return Math.round(dp * getDisplayMetrics(context).density);
 	}
